+++ tests/Tests/PeachApiClient/SearchOffersTests.cs
@@ -9,7 +9,7 @@ public class SearchOffersTests(ITestOutputHelper output)
 {
     const int DEFAULT_DELAY_MS = 1000;
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Search_all_offer_types() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.All,
@@ -19,7 +19,7 @@ public class SearchOffersTests(ITestOutputHelper output)
         MinReputation = 0.5
     });
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Search_ask_offer_types() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.Ask,
@@ -30,7 +30,7 @@ public class SearchOffersTests(ITestOutputHelper output)
     },
     assert: offers => Assert.All(offers, o => Assert.Equal(OfferType.Ask, o.Type)));
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Search_bid_offer_types() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.Bid,
@@ -41,13 +41,13 @@ public class SearchOffersTests(ITestOutputHelper output)
     },
     assert: offers => Assert.All(offers, o => Assert.Equal(OfferType.Bid, o.Type)));
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Search_offers_with_only_mandatory_filters() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.All
     });
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Search_offers_with_pagination() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.All,
@@ -56,7 +56,7 @@ public class SearchOffersTests(ITestOutputHelper output)
     assert: offers => Assert.Single(offers),
     failOnEmpty: true);
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Search_offers_with_sort() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.Ask,
@@ -70,7 +70,7 @@ public class SearchOffersTests(ITestOutputHelper output)
     },
     failOnEmpty: true);
 
-    [Fact, Delay(DEFAULT_DELAY_MS)]
+    [Fact, Delay()]
     public async Task Get_offers_without_pgp_fields() => await SeachOffersAndAssertAsync(new OfferFilter
     {
         Type = OfferTypeFilter.Ask,