+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Refactored registration and authentication methods.
+- Implemented `GetIdentity` method.
+- Implemented `UpdateUserAccount` method.
+- Implemented `CreateOffer` method.
+
 ## [0.1.6-preview] - 2025-09-26
 
 - Implemented registration and authentication methods.
@@ -17,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.4-preview] - 2025-09-06
 
-- Implemented `GetOfferAsync` method. 
+- Implemented `GetOfferAsync` method.
 - Implemented `ValidateResponse` guard method.
 
 ## [0.1.3-preview] - 2025-08-31