+++ CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.4-preview] - 2025-08-31
 
 - Perfected `Offer` type.
 - Implemented `DelayAttribute` to not stress Peach server.
 
-## [0.1.2-previe] - 2025-08-31
+## [0.1.2-preview] - 2025-08-31
 
 - Implemented pagination in `SearchOffersAsync` method.
 - Implemented sort in `SearchOffersAsync` method.